usb: mtp: Improve throughput during multiple file transfer
During multiple file transfers over MTP, the MtpServer
queues an event over the interrupt endpoint before sending
each file. Certain hosts though do not poll on this interrupt
endpoint at all.
This will result in a delay of 1sec for each file transfer
leading to a drastic reduction in throughput.
Fix this by removing the wait for a free request when a event
is queued by the MtpServer.

Change-Id: I75345f3cbff48d1685644f713dbb57a5c772dcf5
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>